+++ frontend/src/hooks/Pagination/useTimeTablePagination.ts
@@ -1,5 +1,5 @@
-import { useRef } from 'react';
-import { useColumnsPerPage } from './useColumnPerPage';
+import { useEffect, useRef } from 'react';
+// import { useColumnsPerPage } from './useColumnPerPage';
 import { usePagination } from './usePagination';
 
 const MIN_COLUMN_WIDTH = 60;
@@ -11,21 +11,33 @@ interface TimeTablePaginationProps {
 export const useTimeTablePagination = ({ availableDates }: TimeTablePaginationProps) => {
   const timeTableContainerRef = useRef<HTMLDivElement>(null);
   const timeColumnRef = useRef<HTMLDivElement>(null);
+  const maxColumnCountPerPageRef = useRef(0);
 
-  const columnCountPerPage = useColumnsPerPage({
-    containerRef: timeTableContainerRef,
-    timeColumnRef,
-    minColumnWidth: MIN_COLUMN_WIDTH,
-  });
+  useEffect(() => {
+    if (!timeTableContainerRef.current || !timeColumnRef.current) return;
+
+    const containerWidth = timeTableContainerRef.current.getBoundingClientRect().width;
+    const style = getComputedStyle(timeTableContainerRef.current);
+    const containerWidthWithoutPadding =
+      containerWidth - parseFloat(style.paddingLeft) - parseFloat(style.paddingRight);
+
+    const timeColumnWidth = timeColumnRef.current.getBoundingClientRect().width;
+
+    maxColumnCountPerPageRef.current = Math.floor(
+      (containerWidthWithoutPadding - timeColumnWidth) / MIN_COLUMN_WIDTH
+    );
+  }, []);
 
   const { totalPages, page, canPagePrev, canPageNext, pagePrev, pageNext, pageReset } =
     usePagination({
       totalItemCount: availableDates.size,
-      perPage: columnCountPerPage,
+      perPage: maxColumnCountPerPageRef.current,
     });
 
-  const startIndex = (page - 1) * columnCountPerPage;
-  const endIndex = startIndex + columnCountPerPage;
+  const maxColumnCountPerPage = maxColumnCountPerPageRef.current;
+
+  const startIndex = (page - 1) * maxColumnCountPerPage;
+  const endIndex = startIndex + maxColumnCountPerPage;
 
   const currentPageDates = new Set(Array.from(availableDates).slice(startIndex, endIndex));
